requirements.txt
Flask==0.10.1
Flask-OAuthlib
Flask-Migrate
Flask-Testing
Flask-Bower
Flask-Assets
Flask-Sqlalchemy==2.1
sqlalchemy-utils==0.31.6
python-dateutil==2.4.2